Move GeoPointMapActivity to geo package
