+++ collect_app/src/main/res/layout/geopoly_layout.xml
@@ -33,7 +33,7 @@
         <androidx.appcompat.widget.AppCompatButton
             android:text="@string/record_geopoint"
             android:id="@+id/record_button"
-            style="@style/Widget.Collect.ImageButton.MapButton"
+            style="@style/Widget.Geo.MapButton"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:padding="15dp"
@@ -62,7 +62,7 @@
 
                 <androidx.appcompat.widget.AppCompatImageButton
                     android:id="@+id/play"
-                    style="@style/Widget.Collect.ImageButton.MapButton"
+                    style="@style/Widget.Geo.MapButton"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:padding="15dp"
@@ -71,7 +71,7 @@
 
                 <androidx.appcompat.widget.AppCompatImageButton
                     android:id="@+id/pause"
-                    style="@style/Widget.Collect.ImageButton.MapButton"
+                    style="@style/Widget.Geo.MapButton"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:padding="15dp"
@@ -81,7 +81,7 @@
 
                 <androidx.appcompat.widget.AppCompatImageButton
                     android:id="@+id/backspace"
-                    style="@style/Widget.Collect.ImageButton.MapButton"
+                    style="@style/Widget.Geo.MapButton"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:padding="15dp"
@@ -90,7 +90,7 @@
 
                 <androidx.appcompat.widget.AppCompatImageButton
                     android:id="@+id/zoom"
-                    style="@style/Widget.Collect.ImageButton.MapButton"
+                    style="@style/Widget.Geo.MapButton"
                     app:srcCompat="@drawable/ic_my_location"
                     android:padding="15dp"
                     android:layout_margin="5dp"
@@ -99,7 +99,7 @@
 
                 <androidx.appcompat.widget.AppCompatImageButton
                     android:id="@+id/layers"
-                    style="@style/Widget.Collect.ImageButton.MapButton"
+                    style="@style/Widget.Geo.MapButton"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:padding="15dp"
@@ -108,7 +108,7 @@
 
                 <androidx.appcompat.widget.AppCompatImageButton
                     android:id="@+id/clear"
-                    style="@style/Widget.Collect.ImageButton.MapButton"
+                    style="@style/Widget.Geo.MapButton"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:padding="15dp"
@@ -117,7 +117,7 @@
 
                 <androidx.appcompat.widget.AppCompatImageButton
                     android:id="@+id/save"
-                    style="@style/Widget.Collect.ImageButton.MapButton"
+                    style="@style/Widget.Geo.MapButton"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:padding="15dp"
