+++ collect_app/src/main/res/layout/instance_map_layout.xml
@@ -36,7 +36,7 @@
 
         <org.odk.collect.android.views.multiclicksafe.MultiClickSafeImageButton
             android:id="@+id/new_instance"
-            style="@style/Widget.Collect.ImageButton.MapButton"
+            style="@style/Widget.Geo.MapButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_margin="5dp"
@@ -59,7 +59,7 @@
 
                 <org.odk.collect.android.views.multiclicksafe.MultiClickSafeImageButton
                     android:id="@+id/zoom_to_location"
-                    style="@style/Widget.Collect.ImageButton.MapButton"
+                    style="@style/Widget.Geo.MapButton"
                     android:layout_width="wrap_content"
                     android:layout_height="match_parent"
                     android:layout_margin="5dp"
@@ -69,7 +69,7 @@
 
                 <org.odk.collect.android.views.multiclicksafe.MultiClickSafeImageButton
                     android:id="@+id/zoom_to_bounds"
-                    style="@style/Widget.Collect.ImageButton.MapButton"
+                    style="@style/Widget.Geo.MapButton"
                     android:layout_width="wrap_content"
                     android:layout_height="match_parent"
                     android:layout_margin="5dp"
@@ -79,7 +79,7 @@
 
                 <androidx.appcompat.widget.AppCompatImageButton
                     android:id="@+id/layer_menu"
-                    style="@style/Widget.Collect.ImageButton.MapButton"
+                    style="@style/Widget.Geo.MapButton"
                     android:layout_width="wrap_content"
                     android:layout_height="match_parent"
                     android:layout_margin="5dp"
