+++ geo/src/main/res/drawable/ic_add_location.xml
