+++ geo/src/main/res/drawable/ic_delete.xml
