+++ geo/src/main/res/drawable/ic_layers.xml
