+++ geo/src/main/res/drawable/ic_my_location.xml
