+++ geo/src/main/res/layout/geopoint_layout.xml
@@ -50,7 +50,7 @@
 
             <androidx.appcompat.widget.AppCompatImageButton
                 android:id="@+id/place_marker"
-                style="@style/Widget.Collect.ImageButton.MapButton"
+                style="@style/Widget.Geo.MapButton"
                 android:layout_width="wrap_content"
                 android:layout_height="match_parent"
                 android:layout_margin="5dp"
@@ -60,7 +60,7 @@
 
             <androidx.appcompat.widget.AppCompatImageButton
                 android:id="@+id/zoom"
-                style="@style/Widget.Collect.ImageButton.MapButton"
+                style="@style/Widget.Geo.MapButton"
                 android:layout_width="wrap_content"
                 android:layout_height="match_parent"
                 android:layout_margin="5dp"
@@ -70,7 +70,7 @@
 
             <androidx.appcompat.widget.AppCompatImageButton
                 android:id="@+id/layer_menu"
-                style="@style/Widget.Collect.ImageButton.MapButton"
+                style="@style/Widget.Geo.MapButton"
                 android:layout_width="wrap_content"
                 android:layout_height="match_parent"
                 android:layout_margin="5dp"
@@ -80,7 +80,7 @@
 
             <androidx.appcompat.widget.AppCompatImageButton
                 android:id="@+id/clear"
-                style="@style/Widget.Collect.ImageButton.MapButton"
+                style="@style/Widget.Geo.MapButton"
                 android:layout_width="wrap_content"
                 android:layout_height="match_parent"
                 android:layout_margin="5dp"
@@ -89,7 +89,7 @@
 
             <androidx.appcompat.widget.AppCompatImageButton
                 android:id="@+id/accept_location"
-                style="@style/Widget.Collect.ImageButton.MapButton"
+                style="@style/Widget.Geo.MapButton"
                 android:layout_width="wrap_content"
                 android:layout_height="match_parent"
                 android:layout_margin="5dp"
